requirements.txt
amazon-braket-default-simulator==1.26.4
amazon-braket-pennylane-plugin==1.31.2
amazon-braket-schemas==1.23.1
amazon-braket-sdk==1.90.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.2
appdirs==1.4.4
astunparse==1.6.3
autograd==1.7.0
autoray==0.7.0
backoff==2.2.1
backports.entry-points-selectable==1.3.0
boltons==25.0.0
boto3==1.37.2
botocore==1.37.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==2.2.1
diastatic-malt==2.15.2
gast==0.6.0
idna==3.10
jmespath==1.0.1
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.0.2
openpulse==1.0.1
openqasm3==1.0.1
opt_einsum==3.4.0
oqpy==0.3.7
packaging==24.2
PennyLane==0.40.0
PennyLane_Lightning==0.40.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
requests==2.32.3
rustworkx==0.16.0
s3transfer==0.11.3
scipy==1.15.2
scipy-openblas32==0.3.29.0.0
six==1.17.0
sympy==1.13.3
termcolor==2.5.0
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
